Rename Rust side
