+++ crates/polars-core/src/series/series_trait.rs
@@ -524,12 +524,12 @@ pub trait SeriesTrait:
     #[cfg(feature = "rolling_window")]
     /// Apply a custom function over a rolling/ moving window of the array.
     /// This has quite some dynamic dispatch, so prefer rolling_min, max, mean, sum over this.
-    fn rolling_apply(
+    fn rolling_map(
         &self,
         _f: &dyn Fn(&Series) -> Series,
         _options: RollingOptionsFixedWindow,
     ) -> PolarsResult<Series> {
-        polars_bail!(opq = rolling_apply, self._dtype());
+        polars_bail!(opq = rolling_map, self._dtype());
     }
     #[cfg(feature = "concat_str")]
     /// Concat the values into a string array.
